move SnapsSettingsList into its own folder
