+++ app/components/Views/Settings/index.js
@@ -9,7 +9,7 @@ import { MetaMetricsEvents } from '../../../core/Analytics';
 import { connect } from 'react-redux';
 import { ThemeContext, mockTheme } from '../../../util/theme';
 import Routes from '../../../constants/navigation/Routes';
-import { createSnapsSettingsListNavDetails } from '../Snaps/SnapsSettingsList';
+import { createSnapsSettingsListNavDetails } from '../Snaps/SnapsSettingsList/SnapsSettingsList';
 
 const createStyles = (colors) =>
   StyleSheet.create({
