+++ app/components/Views/Snaps/SnapsSettingsList/index.ts
